+++ CotEditor/Sources/CEAppDelegate.m
@@ -706,7 +706,7 @@
     
     if (document) {
 //    [document readFromData:[selection dataUsingEncoding:NSUTF8StringEncoding] ofType:nil error:nil];
-        [[document editor] replaceTextViewSelectedStringTo:selection scroll:NO];
+        [[document editor] replaceTextViewSelectedStringWithString:selection];
     } else {
         [[NSAlert alertWithError:err] runModal];
     }