+++ CotEditor/Sources/CEColorCodePanelController.m
@@ -9,7 +9,7 @@
  encoding="UTF-8"
  ------------------------------------------------------------------------------
  
- © 2014 CotEditor Project
+ © 2014-2015 1024jp
  
  This program is free software; you can redistribute it and/or modify it under
  the terms of the GNU General Public License as published by the Free Software
@@ -159,7 +159,10 @@
 - (IBAction)insertCodeToDocument:(id)sender
 // ------------------------------------------------------
 {
-    [[[self documentWindowController] editor] replaceTextViewSelectedStringTo:[self colorCode] scroll:YES];
+    CEEditorWrapper *editor = [[self documentWindowController] editor];
+    
+    [editor replaceTextViewSelectedStringWithString:[self colorCode]];
+    [[editor focusedTextView] scrollRangeToVisible:[[editor focusedTextView] selectedRange]];
 }