+++ CotEditor/Sources/CEToolbarController.m
@@ -10,7 +10,7 @@
  ------------------------------------------------------------------------------
  
  © 2004-2007 nakamuxu
- © 2014 CotEditor Project
+ © 2014-2015 1024jp
  
  This program is free software; you can redistribute it and/or modify it under
  the terms of the GNU General Public License as published by the Free Software
@@ -29,6 +29,7 @@
  */
 
 #import "CEToolbarController.h"
+#import "CEEditorWrapper.h"
 #import "CEEncodingManager.h"
 #import "CESyntaxManager.h"
 #import "CEWindowController.h"